refactor: rename StorageTable to BatchTable
