+++ src/stream/src/common/table/test_storage_table.rs
@@ -24,7 +24,7 @@ use risingwave_common::util::epoch::{test_epoch, EpochPair};
 use risingwave_common::util::sort_util::OrderType;
 use risingwave_hummock_sdk::HummockReadEpoch;
 use risingwave_hummock_test::test_utils::prepare_hummock_test_env;
-use risingwave_storage::table::batch_table::storage_table::StorageTable;
+use risingwave_storage::table::batch_table::BatchTable;
 use risingwave_storage::table::TableIter;
 
 use crate::common::table::state_table::StateTable;
@@ -70,7 +70,7 @@ async fn test_storage_table_value_indices() {
         StateTable::from_table_catalog_inconsistent_op(&table, test_env.storage.clone(), None)
             .await;
 
-    let table = StorageTable::for_test(
+    let table = BatchTable::for_test(
         test_env.storage.clone(),
         TEST_TABLE_ID,
         column_descs.clone(),
@@ -210,7 +210,7 @@ async fn test_shuffled_column_id_for_storage_table_get_row() {
         .start_epoch(epoch.curr, HashSet::from_iter([TEST_TABLE_ID]));
     state.init_epoch(epoch).await.unwrap();
 
-    let table = StorageTable::for_test(
+    let table = BatchTable::for_test(
         test_env.storage.clone(),
         TEST_TABLE_ID,
         column_descs.clone(),
@@ -314,7 +314,7 @@ async fn test_row_based_storage_table_point_get_in_batch_mode() {
             .await;
 
     let column_ids_partial = vec![ColumnId::from(1), ColumnId::from(2)];
-    let table = StorageTable::for_test_with_partial_columns(
+    let table = BatchTable::for_test_with_partial_columns(
         test_env.storage.clone(),
         TEST_TABLE_ID,
         column_descs.clone(),
@@ -425,7 +425,7 @@ async fn test_batch_scan_with_value_indices() {
 
     let column_ids_partial = vec![ColumnId::from(1), ColumnId::from(2)];
 
-    let table = StorageTable::for_test_with_partial_columns(
+    let table = BatchTable::for_test_with_partial_columns(
         test_env.storage.clone(),
         TEST_TABLE_ID,
         column_descs.clone(),
@@ -529,7 +529,7 @@ async fn test_batch_scan_chunk_with_value_indices() {
         .map(|i| ColumnId::from(*i as i32))
         .collect_vec();
 
-    let table = StorageTable::for_test_with_partial_columns(
+    let table = BatchTable::for_test_with_partial_columns(
         test_env.storage.clone(),
         TEST_TABLE_ID,
         column_descs.clone(),
