+++ src/stream/src/executor/backfill/snapshot_backfill/executor.rs
@@ -29,7 +29,7 @@ use risingwave_common::util::epoch::EpochPair;
 use risingwave_common_rate_limit::RateLimit;
 use risingwave_hummock_sdk::HummockReadEpoch;
 use risingwave_storage::store::PrefetchOptions;
-use risingwave_storage::table::batch_table::storage_table::StorageTable;
+use risingwave_storage::table::batch_table::BatchTable;
 use risingwave_storage::table::ChangeLogRow;
 use risingwave_storage::StateStore;
 use tokio::select;
@@ -49,7 +49,7 @@ use crate::task::CreateMviewProgressReporter;
 
 pub struct SnapshotBackfillExecutor<S: StateStore> {
     /// Upstream table
-    upstream_table: StorageTable<S>,
+    upstream_table: BatchTable<S>,
 
     /// Backfill progress table
     progress_state_table: StateTable<S>,
@@ -76,7 +76,7 @@ pub struct SnapshotBackfillExecutor<S: StateStore> {
 impl<S: StateStore> SnapshotBackfillExecutor<S> {
     #[expect(clippy::too_many_arguments)]
     pub(crate) fn new(
-        upstream_table: StorageTable<S>,
+        upstream_table: BatchTable<S>,
         progress_state_table: StateTable<S>,
         upstream: MergeExecutorInput,
         output_indices: Vec<usize>,
@@ -570,7 +570,7 @@ async fn receive_next_barrier(
 }
 
 async fn make_log_stream(
-    upstream_table: &StorageTable<impl StateStore>,
+    upstream_table: &BatchTable<impl StateStore>,
     prev_epoch: u64,
     start_pk: Option<OwnedRow>,
     chunk_size: usize,
@@ -601,7 +601,7 @@ async fn make_log_stream(
 }
 
 async fn make_snapshot_stream(
-    upstream_table: &StorageTable<impl StateStore>,
+    upstream_table: &BatchTable<impl StateStore>,
     snapshot_epoch: u64,
     start_pk: Option<OwnedRow>,
     rate_limit: RateLimit,
@@ -634,7 +634,7 @@ async fn make_snapshot_stream(
 #[expect(clippy::too_many_arguments)]
 #[try_stream(ok = Message, error = StreamExecutorError)]
 async fn make_consume_snapshot_stream<'a, S: StateStore>(
-    upstream_table: &'a StorageTable<S>,
+    upstream_table: &'a BatchTable<S>,
     snapshot_epoch: u64,
     chunk_size: usize,
     rate_limit: RateLimit,
