+++ src/stream/src/executor/lookup/impl_.rs
@@ -21,7 +21,7 @@ use risingwave_common::util::sort_util::ColumnOrder;
 use risingwave_common_estimate_size::collections::EstimatedVec;
 use risingwave_hummock_sdk::HummockReadEpoch;
 use risingwave_storage::store::PrefetchOptions;
-use risingwave_storage::table::batch_table::storage_table::StorageTable;
+use risingwave_storage::table::batch_table::BatchTable;
 use risingwave_storage::table::TableIter;
 
 use super::sides::{stream_lookup_arrange_prev_epoch, stream_lookup_arrange_this_epoch};
@@ -89,7 +89,7 @@ pub struct LookupExecutorParams<S: StateStore> {
     /// The join keys on the arrangement side.
     pub arrange_join_key_indices: Vec<usize>,
 
-    pub storage_table: StorageTable<S>,
+    pub batch_table: BatchTable<S>,
 
     pub watermark_epoch: AtomicU64Ref,
 
@@ -109,7 +109,7 @@ impl<S: StateStore> LookupExecutor<S> {
             stream_join_key_indices,
             arrange_join_key_indices,
             column_mapping,
-            storage_table,
+            batch_table: storage_table,
             watermark_epoch,
             chunk_size,
         } = params;
@@ -209,7 +209,7 @@ impl<S: StateStore> LookupExecutor<S> {
                 order_rules: arrangement_order_rules,
                 key_indices: arrange_join_key_indices,
                 use_current_epoch,
-                storage_table,
+                batch_table: storage_table,
             },
             column_mapping,
             key_indices_mapping,
@@ -240,7 +240,7 @@ impl<S: StateStore> LookupExecutor<S> {
         };
 
         let metrics = self.ctx.streaming_metrics.new_lookup_executor_metrics(
-            self.arrangement.storage_table.table_id(),
+            self.arrangement.batch_table.table_id(),
             self.ctx.id,
             self.ctx.fragment_id,
         );
@@ -329,7 +329,7 @@ impl<S: StateStore> LookupExecutor<S> {
         if let Some(vnode_bitmap) = barrier.as_update_vnode_bitmap(self.ctx.id) {
             let previous_vnode_bitmap = self
                 .arrangement
-                .storage_table
+                .batch_table
                 .update_vnode_bitmap(vnode_bitmap.clone());
 
             // Manipulate the cache if necessary.
@@ -371,7 +371,7 @@ impl<S: StateStore> LookupExecutor<S> {
             let all_data_iter = match self.arrangement.use_current_epoch {
                 true => {
                     self.arrangement
-                        .storage_table
+                        .batch_table
                         .batch_iter_with_pk_bounds(
                             HummockReadEpoch::NoWait(epoch_pair.curr),
                             &lookup_row,
@@ -383,7 +383,7 @@ impl<S: StateStore> LookupExecutor<S> {
                 }
                 false => {
                     self.arrangement
-                        .storage_table
+                        .batch_table
                         .batch_iter_with_pk_bounds(
                             HummockReadEpoch::NoWait(epoch_pair.prev),
                             &lookup_row,
