+++ src/stream/src/executor/lookup/sides.rs
@@ -24,7 +24,7 @@ use risingwave_common::bail;
 use risingwave_common::catalog::ColumnDesc;
 use risingwave_common::types::DataType;
 use risingwave_common::util::sort_util::ColumnOrder;
-use risingwave_storage::table::batch_table::storage_table::StorageTable;
+use risingwave_storage::table::batch_table::BatchTable;
 use risingwave_storage::StateStore;
 
 use crate::executor::error::StreamExecutorError;
@@ -76,7 +76,7 @@ pub(crate) struct ArrangeJoinSide<S: StateStore> {
     /// Whether to join with the arrangement of the current epoch
     pub use_current_epoch: bool,
 
-    pub storage_table: StorageTable<S>,
+    pub batch_table: BatchTable<S>,
 }
 
 /// Message from the `arrange_join_stream`.
