+++ src/stream/src/from_proto/batch_query.rs
@@ -15,7 +15,7 @@
 use risingwave_common::catalog::ColumnId;
 use risingwave_pb::plan_common::StorageTableDesc;
 use risingwave_pb::stream_plan::BatchPlanNode;
-use risingwave_storage::table::batch_table::storage_table::StorageTable;
+use risingwave_storage::table::batch_table::BatchTable;
 
 use super::*;
 use crate::executor::{BatchQueryExecutor, DummyExecutor};
@@ -46,7 +46,7 @@ impl ExecutorBuilder for BatchQueryExecutorBuilder {
             .map(ColumnId::from)
             .collect();
 
-        let table = StorageTable::new_partial(
+        let table = BatchTable::new_partial(
             state_store,
             column_ids,
             params.vnode_bitmap.map(Into::into),
