+++ src/stream/src/from_proto/stream_scan.rs
@@ -19,7 +19,7 @@ use risingwave_common::util::value_encoding::column_aware_row_encoding::ColumnAw
 use risingwave_common::util::value_encoding::BasicSerde;
 use risingwave_pb::plan_common::StorageTableDesc;
 use risingwave_pb::stream_plan::{StreamScanNode, StreamScanType};
-use risingwave_storage::table::batch_table::storage_table::StorageTable;
+use risingwave_storage::table::batch_table::BatchTable;
 
 use super::*;
 use crate::common::table::state_table::{ReplicatedStateTable, StateTable};
@@ -80,7 +80,7 @@ impl ExecutorBuilder for StreamScanExecutorBuilder {
                 };
 
                 let upstream_table =
-                    StorageTable::new_partial(state_store.clone(), column_ids, vnodes, table_desc);
+                    BatchTable::new_partial(state_store.clone(), column_ids, vnodes, table_desc);
 
                 BackfillExecutor::new(
                     upstream_table,
