+++ src/stream/src/from_proto/watermark_filter.rs
@@ -17,7 +17,7 @@ use std::sync::Arc;
 use risingwave_common::catalog::{ColumnId, TableDesc};
 use risingwave_expr::expr::build_non_strict_from_prost;
 use risingwave_pb::stream_plan::WatermarkFilterNode;
-use risingwave_storage::table::batch_table::storage_table::StorageTable;
+use risingwave_storage::table::batch_table::BatchTable;
 
 use super::*;
 use crate::common::table::state_table::StateTable;
@@ -57,7 +57,7 @@ impl ExecutorBuilder for WatermarkFilterBuilder {
             .collect_vec();
         let other_vnodes = Arc::new(!(*vnodes).clone());
         let global_watermark_table =
-            StorageTable::new_partial(store.clone(), column_ids, Some(other_vnodes), &desc);
+            BatchTable::new_partial(store.clone(), column_ids, Some(other_vnodes), &desc);
 
         let table =
             StateTable::from_table_catalog_inconsistent_op(&table, store, Some(vnodes)).await;
